* A crazy amount of changes.  * The ckFileSystem core should be close to stable along with the GUI    associated with it.

+++ Source/Core/ckFileSystem/Joliet.cpp
@@ -337,7 +337,7 @@ namespace ckFileSystem
 		return m_iMaxNameLen;*/
 
 		size_t iNameLen = lstrlen(szFileName);
-		if (iNameLen >= m_iMaxNameLen)
+		if (iNameLen >= (size_t)m_iMaxNameLen)
 			iNameLen = m_iMaxNameLen;
 
 		if (!bIsDir && m_bIncFileVerInfo)
